Prevented "--all" from being passed to git-shortlog.
When it is passed to git-shortlog, "--all" tells it to include commits
from all branches, not just the current one.  While this might be a nice
feature to have, it should probably be a separate option rather than
being an overloaded meaning of a single option that also just enables
per-contributor counts.