Construct commit message incrementally as you work
This adds the `git commit-message` command with actions: add,
edit, clear, and show (the default).  You can use it to build
up a commit message while you are working, simply running:

    git commit-message add A quick reminder

When you eventually do the `git commit`, the message that you
have accumulated will be the seed for the commit message to be
used.

Note that this sets up prepare-commit-msg and post-commit hooks
for you when used so that it can inject the commit message and
clear it out after the commit.