Added automated manual generation.
Bash script which rebuilds the indexes (index.txt and the links listed in git-extras.md) based on the collection of .md file documents avalable.
By interrogating the .md file for the documented description we null and void yet another possible consistency issue and avoid manual duplication mistakes.
Once the indexes are up to date we start from the begining of the directory listing again calling ronn on each .md to generate the man and html files as well as renaming the incerrectly name html file.

Result several incorrectly named .html documents are now removed, and all the documentation is up to date.
Documentation also updated.