FIXes 'mktemp' to work again on Linux
Broken in 67241a46a820f19706a398adfee55af863fb5449

`mktemp` for BSD (including OSX) requires a template, but it allows any number
of `X`s in the template.

(GNU) `mktemp` for Linux does not require a template, however, if a template is
specified, then the number of `X`s must be 6.

See http://unix.stackexchange.com/a/206111/117157