feat(git-release): push tags first to help ci tools not miss them
If you are buliding on CircleCI for example, pushing the commits first
can lead to a race where CI env will not see the tags since the pull
happened before the tags were finished pushing.

It should be fully safe (and backwards-compatible) to change the order
of pushing to the remote.